components: migrate ArtistsColumn and RowHeader to typescript
