+++ src/types.ts
@@ -1,0 +1,10 @@
+export type Artist = {
+    id: number;
+    name: string;
+    slug: string;
+    show_count: number;
+    source_count: number;
+    featured: number;
+    sort_name: string;
+    the: boolean;
+}
